requirements.txt
matplotlib
PyYAML
huggingface_hub
datasets
transformers
torchvision
scikit-learn
accelerate>=0.26.0
